fix log viewer
