+++ src/packages/log-viewer/components/log-viewer-date-range-selector.element.ts
@@ -1,5 +1,5 @@
 import type { LogViewerDateRange, UmbLogViewerWorkspaceContext } from '../workspace/logviewer.context.js';
-import { UMB_APP_LOG_VIEWER_CONTEXT } from '../workspace/logviewer.context-token.js';
+import { UMB_APP_LOG_VIEWER_CONTEXT } from '../workspace/logviewer-workspace.context-token.js';
 import { UmbTextStyles } from '@umbraco-cms/backoffice/style';
 import { css, html, customElement, property, queryAll, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
