+++ src/packages/log-viewer/workspace/logviewer-workspace.context-token.ts
