+++ src/packages/log-viewer/workspace/manifests.ts
@@ -1,3 +1,71 @@
-import { manifests as logViewerRootManifests } from './logviewer/manifests.js';
+import type {
+	ManifestModal,
+	ManifestWorkspace,
+	ManifestWorkspaceActions,
+	ManifestWorkspaceView,
+} from '@umbraco-cms/backoffice/extension-registry';
 
-export const manifests = [...logViewerRootManifests];
+const workspaceAlias = 'Umb.Workspace.LogViewer';
+
+const workspace: ManifestWorkspace = {
+	type: 'workspace',
+	alias: workspaceAlias,
+	name: 'LogViewer Root Workspace',
+	element: () => import('./logviewer-workspace.element.js'),
+	api: () => import('./logviewer-workspace.context.js'),
+	meta: {
+		entityType: 'logviewer',
+	},
+};
+
+const workspaceViews: Array<ManifestWorkspaceView> = [
+	{
+		type: 'workspaceView',
+		alias: 'Umb.WorkspaceView.LogViewer.Overview',
+		name: 'LogViewer Root Workspace Overview View',
+		element: () => import('./views/overview/index.js'),
+		weight: 300,
+		meta: {
+			label: 'Overview',
+			pathname: 'overview',
+			icon: 'icon-box-alt',
+		},
+		conditions: [
+			{
+				alias: 'Umb.Condition.WorkspaceAlias',
+				match: workspace.alias,
+			},
+		],
+	},
+	{
+		type: 'workspaceView',
+		alias: 'Umb.WorkspaceView.LogViewer.Search',
+		name: 'LogViewer Root Workspace Search View',
+		element: () => import('./views/search/index.js'),
+		weight: 200,
+		meta: {
+			label: 'Search',
+			pathname: 'search',
+			icon: 'icon-search',
+		},
+		conditions: [
+			{
+				alias: 'Umb.Condition.WorkspaceAlias',
+				match: workspace.alias,
+			},
+		],
+	},
+];
+
+const workspaceActions: Array<ManifestWorkspaceActions> = [];
+
+const modals: Array<ManifestModal> = [
+	{
+		type: 'modal',
+		alias: 'Umb.Modal.LogViewer.SaveSearch',
+		name: 'Saved Searches Modal',
+		element: () => import('./views/search/components/log-viewer-search-input-modal.element.js'),
+	},
+];
+
+export const manifests = [workspace, ...workspaceViews, ...workspaceActions, ...modals];
