+++ src/packages/log-viewer/workspace/views/overview/log-overview-view.element.ts
@@ -1,5 +1,5 @@
 import type { UmbLogViewerWorkspaceContext } from '../../logviewer.context.js';
-import { UMB_APP_LOG_VIEWER_CONTEXT } from '../../logviewer.context-token.js';
+import { UMB_APP_LOG_VIEWER_CONTEXT } from '../../logviewer-workspace.context-token.js';
 import { css, html, customElement, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
