+++ src/packages/log-viewer/workspace/views/search/components/log-viewer-polling-button.element.ts
@@ -1,5 +1,5 @@
 import type { PoolingCOnfig, PoolingInterval, UmbLogViewerWorkspaceContext } from '../../../logviewer.context.js';
-import { UMB_APP_LOG_VIEWER_CONTEXT } from '../../../logviewer.context-token.js';
+import { UMB_APP_LOG_VIEWER_CONTEXT } from '../../../logviewer-workspace.context-token.js';
 import { css, html, customElement, query, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import type { UmbDropdownElement } from '@umbraco-cms/backoffice/components';
