+++ src/packages/log-viewer/workspace/views/search/components/log-viewer-search-input.element.ts
@@ -1,5 +1,5 @@
 import type { UmbLogViewerWorkspaceContext } from '../../../logviewer.context.js';
-import { UMB_APP_LOG_VIEWER_CONTEXT } from '../../../logviewer.context-token.js';
+import { UMB_APP_LOG_VIEWER_CONTEXT } from '../../../logviewer-workspace.context-token.js';
 import type {
 	UmbContextSaveSearchModalData,
 	UmbContextSaveSearchModalValue,
